// app/lib/types.ts
export interface UIDaily {
    date: string; // YYYY-MM-DD
    dayName: string; // "Mon", "Tue"
    max_c: number;
    max_f: number;
    min_c: number;
    min_f: number;
    conditionText: string;
    conditionIcon: string;
    chance_of_rain: number;
    isMock?: boolean;
}